Refactor error modeling in cxi.merge by moving sdfac_refine and errors_from_sample_residuals into separate class files
Add xfel/merging/algorithms/error_model folder.
Add base class for error modeling: merging/algorithms/error_model/error_modeler_base.
Move sdfac_refine and errors_from_sample_residuals code into classes in error_model folder.
Adjust cxi.merge, cxi.xmerge and dev.cxi.mpi_cluster_two_merge to use these new classes.
